+++ app/controllers/replies_controller.rb
@@ -104,7 +104,7 @@ class RepliesController < ApplicationController
   end
 
   def redirect_without_reply
-    empty_response_or_root_path(@chapter) unless @reply
+    empty_response_or_root_path(@chapter || chapters_path) unless @reply
   end
 
   def reply_params